+++ services/ui-src/src/routes/McparReportPage/McparReportPage.tsx
@@ -1,41 +1,25 @@
-import { useContext, useState } from "react";
+import { useContext } from "react";
 import { useNavigate } from "react-router-dom";
 // components
-import { Box, Button, Flex, Heading, useDisclosure } from "@chakra-ui/react";
+import { Flex } from "@chakra-ui/react";
 import {
-  Form,
-  Icon,
   ReportContext,
-  ReportDrawer,
   ReportPage,
+  ReportPageEntitySection,
+  ReportPageFormSection,
+  ReportPageIntro,
   Sidebar,
-  SpreadsheetWidget,
 } from "components";
 // utils
-import { findRoute, hydrateFormFields, parseCustomHtml, useUser } from "utils";
-import {
-  AnyObject,
-  CustomHtmlElement,
-  ReportStatus,
-  SpreadsheetWidgetProps,
-  UserRoles,
-} from "types";
+import { findRoute, useUser } from "utils";
+import { PageJson, ReportStatus, UserRoles } from "types";
 // form data
 import { mcparRoutes } from "forms/mcpar";
-import { reportSchema } from "forms/mcpar/reportSchema";
 
 export const McparReportPage = ({ pageJson }: Props) => {
   const navigate = useNavigate();
-  const { reportData, updateReportData, updateReport } =
-    useContext(ReportContext);
-  const { path, pageType, intro, form, drawer } = pageJson;
-  const { isOpen, onClose, onOpen } = useDisclosure();
-
-  // make state
-  const [currentEntity, setCurrentEntity] = useState("");
-
-  // make routes
-  const previousRoute = findRoute(mcparRoutes, path, "previous", "/mcpar");
+  const { updateReportData, updateReport } = useContext(ReportContext);
+  const { path, pageType, intro, form } = pageJson;
   const nextRoute = findRoute(mcparRoutes, path, "next", "/mcpar");
 
   // get user's state
@@ -58,17 +42,17 @@ export const McparReportPage = ({ pageJson }: Props) => {
     navigate(nextRoute);
   };
 
-  const openRowDrawer = (entity: string) => {
-    setCurrentEntity(entity);
-    onOpen();
-  };
-
-  if (reportData) {
-    form.fields = hydrateFormFields(form.fields, reportData);
-  }
-
-  const tempEntityMap = {
-    plans: ["United Healthcare", "Care 1st", "Aetna Family Care"],
+  const renderPageSection = (pageType: string) => {
+    switch (pageType) {
+      case "drawer":
+        return (
+          <ReportPageEntitySection pageJson={pageJson} onSubmit={onSubmit} />
+        );
+      default:
+        return (
+          <ReportPageFormSection pageJson={pageJson} onSubmit={onSubmit} />
+        );
+    }
   };
 
   return (
@@ -76,55 +60,8 @@ export const McparReportPage = ({ pageJson }: Props) => {
       <Flex sx={sx.pageContainer}>
         <Sidebar />
         <Flex sx={sx.reportContainer}>
-          <ReportPageIntro text={intro} />
-          {pageType === "drawer" ? (
-            <>
-              <Box>
-                <Heading as="h3">{drawer.dashboard.title}</Heading>
-                {tempEntityMap.plans.map((entity) => {
-                  return (
-                    <Flex key={entity} sx={sx.entityRow}>
-                      <Heading as="h4">{entity}</Heading>
-                      <Button
-                        sx={sx.enterButton}
-                        onClick={() => openRowDrawer(entity)}
-                        variant="outline"
-                      >
-                        Enter
-                      </Button>
-                    </Flex>
-                  );
-                })}
-              </Box>
-              <ReportDrawer
-                drawerDisclosure={{
-                  isOpen,
-                  onClose,
-                }}
-                drawerTitle={`${drawer.drawerTitle} ${currentEntity}`}
-                form={form}
-                onSubmit={onSubmit}
-              />
-              <ReportPageFooter
-                previousRoute={previousRoute}
-                nextRoute={nextRoute}
-              />
-            </>
-          ) : (
-            <>
-              <Form
-                id={form.id}
-                formJson={form}
-                formSchema={reportSchema[form.id as keyof typeof reportSchema]}
-                onSubmit={onSubmit}
-              />
-              <ReportPageFooter
-                formId={form.id}
-                previousRoute={previousRoute}
-                nextRoute={nextRoute}
-              />
-            </>
-          )}
+          {intro && <ReportPageIntro text={intro} />}
+          {renderPageSection(pageType)}
         </Flex>
       </Flex>
     </ReportPage>
@@ -132,82 +69,7 @@ export const McparReportPage = ({ pageJson }: Props) => {
 };
 
 interface Props {
-  pageJson: AnyObject;
-}
-
-const ReportPageIntro = ({ text }: ReportPageIntroI) => {
-  const { section, subsection, info, spreadsheet } = text;
-  return (
-    <Box sx={sx.introBox}>
-      <Heading as="h1" sx={sx.sectionHeading}>
-        {section}
-      </Heading>
-      <Heading as="h2" sx={sx.subsectionHeading}>
-        {subsection}
-      </Heading>
-      {spreadsheet && (
-        <Box sx={sx.spreadsheetWidgetBox}>
-          <SpreadsheetWidget content={spreadsheet} />
-        </Box>
-      )}
-      {info && <Box sx={sx.infoTextBox}>{parseCustomHtml(info)}</Box>}
-    </Box>
-  );
-};
-
-interface ReportPageIntroI {
-  text: {
-    section: string;
-    subsection: string;
-    info?: CustomHtmlElement[];
-    spreadsheet?: SpreadsheetWidgetProps;
-  };
-}
-
-const ReportPageFooter = ({
-  formId,
-  previousRoute,
-  nextRoute,
-}: ReportPageFooterI) => {
-  const navigate = useNavigate();
-  return (
-    <Box sx={sx.footerBox}>
-      <Box>
-        <Flex sx={sx.buttonFlex}>
-          <Button
-            onClick={() => navigate(previousRoute)}
-            variant="outline"
-            leftIcon={<Icon icon="arrowLeft" />}
-          >
-            Previous
-          </Button>
-          {formId ? (
-            <Button
-              form={formId}
-              type="submit"
-              rightIcon={<Icon icon="arrowRight" />}
-            >
-              Save & continue
-            </Button>
-          ) : (
-            <Button
-              onClick={() => navigate(nextRoute)}
-              rightIcon={<Icon icon="arrowRight" />}
-            >
-              Continue
-            </Button>
-          )}
-        </Flex>
-        {/* TODO: Add Prince Print Button */}
-      </Box>
-    </Box>
-  );
-};
-
-interface ReportPageFooterI {
-  formId?: string;
-  previousRoute: string;
-  nextRoute: string;
+  pageJson: PageJson;
 }
 
 const sx = {
@@ -221,68 +83,16 @@ const sx = {
     maxWidth: "reportPageWidth",
     marginY: "3.5rem",
     marginLeft: "3.5rem",
-    h4: {
-      fontSize: "lg",
-      fontWeight: "bold",
-    },
     h3: {
-      fontSize: "lg",
-      color: "palette.gray_medium",
-      fontWeight: "bold",
       paddingBottom: "0.75rem",
       borderBottom: "1.5px solid var(--chakra-colors-palette-gray_lighter)",
+      color: "palette.gray_medium",
+      fontSize: "lg",
+      fontWeight: "bold",
     },
-  },
-  introBox: {
-    marginBottom: "2rem",
-  },
-  sectionHeading: {
-    color: "palette.gray",
-    fontSize: "md",
-  },
-  subsectionHeading: {
-    fontWeight: "normal",
-    fontSize: "4xl",
-  },
-  infoTextBox: {
-    marginTop: "2rem",
     h4: {
       fontSize: "lg",
-      marginBottom: "0.75rem",
+      fontWeight: "bold",
     },
-    "p, span": {
-      color: "palette.gray",
-    },
-    a: {
-      color: "palette.primary",
-      "&:hover": {
-        color: "palette.primary_darker",
-      },
-    },
-  },
-  enterButton: {
-    fontSize: "sm",
-    fontWeight: "normal",
-    height: "1.75rem",
-    width: "4.25rem",
-  },
-  entityRow: {
-    justifyContent: "space-between",
-    height: "3.25rem",
-    padding: "0.5rem",
-    paddingLeft: "0.75rem",
-    alignItems: "center",
-    borderBottom: "1.5px solid var(--chakra-colors-palette-gray_lighter)",
-  },
-  spreadsheetWidgetBox: {
-    marginTop: "2rem",
-  },
-  footerBox: {
-    marginTop: "3.5rem",
-    borderTop: "1.5px solid var(--chakra-colors-palette-gray_light)",
-  },
-  buttonFlex: {
-    justifyContent: "space-between",
-    marginY: "1.5rem",
   },
 };
